+++ exporters/prometheus/exporter_test.go
@@ -26,6 +26,7 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	otelmetric "go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/instrument"
+	"go.opentelemetry.io/otel/metric/unit"
 	"go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/view"
@@ -39,7 +40,7 @@ func TestPrometheusExporter(t *testing.T) {
 		emptyResource      bool
 		customResouceAttrs []attribute.KeyValue
 		recordMetrics      func(ctx context.Context, meter otelmetric.Meter)
-		withoutTargetInfo  bool
+		options            []Option
 		expectedFile       string
 	}{
 		{
@@ -52,7 +53,11 @@ func TestPrometheusExporter(t *testing.T) {
 					attribute.Key("E").Bool(true),
 					attribute.Key("F").Int(42),
 				}
-				counter, err := meter.SyncFloat64().Counter("foo", instrument.WithDescription("a simple counter"))
+				counter, err := meter.SyncFloat64().Counter(
+					"foo",
+					instrument.WithDescription("a simple counter"),
+					instrument.WithUnit(unit.Milliseconds),
+				)
 				require.NoError(t, err)
 				counter.Add(ctx, 5, attrs...)
 				counter.Add(ctx, 10.3, attrs...)
@@ -67,10 +72,14 @@ func TestPrometheusExporter(t *testing.T) {
 					attribute.Key("A").String("B"),
 					attribute.Key("C").String("D"),
 				}
-				gauge, err := meter.SyncFloat64().UpDownCounter("bar", instrument.WithDescription("a fun little gauge"))
+				gauge, err := meter.SyncFloat64().UpDownCounter(
+					"bar",
+					instrument.WithDescription("a fun little gauge"),
+					instrument.WithUnit(unit.Dimensionless),
+				)
 				require.NoError(t, err)
-				gauge.Add(ctx, 100, attrs...)
-				gauge.Add(ctx, -25, attrs...)
+				gauge.Add(ctx, 1.0, attrs...)
+				gauge.Add(ctx, -.25, attrs...)
 			},
 		},
 		{
@@ -81,7 +90,11 @@ func TestPrometheusExporter(t *testing.T) {
 					attribute.Key("A").String("B"),
 					attribute.Key("C").String("D"),
 				}
-				histogram, err := meter.SyncFloat64().Histogram("histogram_baz", instrument.WithDescription("a very nice histogram"))
+				histogram, err := meter.SyncFloat64().Histogram(
+					"histogram_baz",
+					instrument.WithDescription("a very nice histogram"),
+					instrument.WithUnit(unit.Bytes),
+				)
 				require.NoError(t, err)
 				histogram.Record(ctx, 23, attrs...)
 				histogram.Record(ctx, 7, attrs...)
@@ -92,6 +105,7 @@ func TestPrometheusExporter(t *testing.T) {
 		{
 			name:         "sanitized attributes to labels",
 			expectedFile: "testdata/sanitized_labels.txt",
+			options:      []Option{WithoutUnits()},
 			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
 				attrs := []attribute.KeyValue{
 					// exact match, value should be overwritten
@@ -102,7 +116,12 @@ func TestPrometheusExporter(t *testing.T) {
 					attribute.Key("C.D").String("Y"),
 					attribute.Key("C/D").String("Z"),
 				}
-				counter, err := meter.SyncFloat64().Counter("foo", instrument.WithDescription("a sanitary counter"))
+				counter, err := meter.SyncFloat64().Counter(
+					"foo",
+					instrument.WithDescription("a sanitary counter"),
+					// This unit is not added to
+					instrument.WithUnit(unit.Bytes),
+				)
 				require.NoError(t, err)
 				counter.Add(ctx, 5, attrs...)
 				counter.Add(ctx, 10.3, attrs...)
@@ -177,9 +196,9 @@ func TestPrometheusExporter(t *testing.T) {
 			},
 		},
 		{
-			name:              "without target_info",
-			withoutTargetInfo: true,
-			expectedFile:      "testdata/without_target_info.txt",
+			name:         "without target_info",
+			options:      []Option{WithoutTargetInfo()},
+			expectedFile: "testdata/without_target_info.txt",
 			recordMetrics: func(ctx context.Context, meter otelmetric.Meter) {
 				attrs := []attribute.KeyValue{
 					attribute.Key("A").String("B"),
@@ -200,13 +219,7 @@ func TestPrometheusExporter(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			ctx := context.Background()
 			registry := prometheus.NewRegistry()
-
-			opts := []Option{WithRegisterer(registry)}
-			if tc.withoutTargetInfo {
-				opts = append(opts, WithoutTargetInfo())
-			}
-
-			exporter, err := New(opts...)
+			exporter, err := New(append(tc.options, WithRegisterer(registry))...)
 			require.NoError(t, err)
 
 			customBucketsView, err := view.New(